requirements.txt
aiohttp==3.9.1
aioserial==1.3.1
devtools==0.12.2
fastapi==0.105.0
gpiozero==1.6.2
httpx==0.26.0
Jinja2==3.1.2
luma.oled==3.13.0
more_itertools==10.1.0
opencv-contrib-python==4.9.0.80
polling2==0.5.0
protobuf==3.20.3
pydantic==2.5.2
pydantic_settings==2.1.0
pyserial==3.5
py-spy==0.3.14
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
sentry-sdk==1.39.1
typer[all]==0.9.0
uvicorn[standard]==0.24.0.post1
v4l2py==2.3.0
watchfiles==0.21.0
websocket-client==1.7.0
ydlidar @ git+https://git@github.com/YDLIDAR/YDLidar-SDK.git@V1.1.3#egg=ydlidar
